_Object:: Make debug_oid uint64.

+++ CoreLibrary/base.cpp
@@ -84,7 +84,7 @@ namespace core {
 
 #ifdef WITH_DEBUG_OID
 // Start with a non-zero value so that it doesn't appear to track object OIDs.
-static uint32 last_debug_oid = 10;
+static uint64 last_debug_oid = 10;
 #endif
 
 _Object::_Object() : refCount_(0) {
@@ -99,7 +99,7 @@ _Object::~_Object() {
 }
 
 #ifdef WITH_DEBUG_OID
-void _Object::set_debug_oid(uint32 debug_oid) {
+void _Object::set_debug_oid(uint64 debug_oid) {
   debug_oid_ = debug_oid;
   // Make sure the next assigned debug OID is higher.
   last_debug_oid = debug_oid;
